refactor(correlation): share the mint/adopt mechanism, keep the namespace per context
The two correlation modules were byte-identical apart from one line
naming their own context — held that way by a boundary test that read
both files and compared the strings. That test was the evidence: a
duplication forbidden to diverge is a shared kernel paying duplication
cost without the one-fix-site benefit.

The identity mechanism moves to @music/eventing: mint, adopt-vs-mint,
causation chaining, the format, the degrade rules, and the parser for
metadata no upcaster protects. Each context keeps the one fact that is
genuinely its own — which store its derived coordinates address — plus
its own logger type through a scope.

This amends end-to-end-correlation D13 rather than overturning it: the
envelope still crosses the seam as data, the facades still take a plain
StoryId string, and no model type is shared. What changed is its
conclusion that the module had to be duplicated; an opaque observability
identity carries neither context's language.

Also records the amendment where a future reader will look: the
CONTEXT-MAP now says no shared MODEL and names the checkpointed drain,
and the three superseded decisions carry annotations pointing here —
including the drifted Marten citation behind the old park-or-halt pair.
