+++ openspec/changes/archive/2026-08-07-end-to-end-correlation/design.md
@@ -129,6 +129,15 @@ Alternative considered and deferred: making every command against an existing st
 stream's story. That is a broader semantic change (it would supersede the request's own mint for
 cancel/select/resolve) and was out of scope here.
 
+> **Amended by `extract-eventing-package` D5 (2026-08-18).** The mint/adopt/causation *mechanism*
+> now lives once in `packages/eventing`; each context keeps only the name it stamps onto derived
+> coordinates. This decision's core claim is untouched — the envelope still crosses the seam as
+> DATA, the facades still take a plain `StoryId` string, and no model type is shared. What changed
+> is its conclusion that the MODULE had to be duplicated: an opaque observability identity carries
+> neither context's language, and the boundary test that held the two copies byte-identical was
+> itself the evidence (a shared kernel without its benefit — see
+> `docs/research/bounded-contexts-vs-modules.md` §pitfalls).
+
 **D13 — The pair is duplicated per context, not shared.** Each module owns its own
 `CommandContext`/`CausationReference`/`CorrelationSource`, exactly as each already owns a
 byte-identical `EventStorePort` and `EventMetadata`. A shared correlation module would be a shared
