+++ packages/importer/src/application/correlation/context.test.ts
@@ -1,198 +1,59 @@
 import { describe, expect, it } from 'vitest';
-import type { StoredEvent } from '../ports/event-store-port.js';
-import {
-  CONTEXT_NAME,
-  adoptOrMint,
-  parseCausation,
-  adoptStory,
-  causedBy,
-  continueFrom,
-  isCorrelationId,
-  newOperation,
-  toCorrelationId,
-} from './context.js';
-import type { CorrelationSource } from '../ports/system-ports.js';
+import { CONTEXT_NAME, continueFrom, newOperation, operationScope } from './context.js';
+import { fixedCorrelation, STORY, OTHER_STORY } from '../__fixtures__/correlation.js';
+import { createLogger } from '../logging/logger.js';
+import type { DestinationStream } from 'pino';
 
-const source = (...values: readonly string[]): CorrelationSource => {
-  let index = 0;
-  return {
-    mint: () => toCorrelationId(values[Math.min(index++, values.length - 1)]!),
-  };
-};
+/**
+ * The mechanism — mint, adopt, the causation parser, the degrade rules — belongs to
+ * `@music/eventing` and is tested there. What this module adds is the BINDING: which store its
+ * derived coordinates address, and this module's own logger type carried through a scope. Only
+ * those two facts are asserted here.
+ */
 
-const STORY = 'a1b2c3d4e5f60718293a4b5c6d7e8f90';
-const OTHER = '00112233445566778899aabbccddeeff';
-
-const storedEvent = (metadata: StoredEvent['metadata']): StoredEvent =>
-  ({
-    globalSeq: 7,
-    streamId: 'imp-1',
-    version: 3,
-    type: 'ImportSubmitted',
-    event: { type: 'ImportSubmitted' },
-    metadata,
-  }) as unknown as StoredEvent;
-
-describe('isCorrelationId', () => {
-  it('accepts a 32-character lowercase hex string', () => {
-    expect(isCorrelationId(STORY)).toBe(true);
-  });
-
-  it.each([
-    ['too short', 'a1b2c3'],
-    ['too long', `${STORY}00`],
-    ['uppercase hex', STORY.toUpperCase()],
-    ['non-hex characters', 'z1b2c3d4e5f60718293a4b5c6d7e8f90'],
-    ['a uuid with dashes', 'a1b2c3d4-e5f6-0718-293a-4b5c6d7e8f90'],
-    ['empty', ''],
-  ])('rejects %s', (_label, value) => {
-    expect(isCorrelationId(value)).toBe(false);
-  });
+const triggeringEvent = (correlationId?: string) => ({
+  streamId: 'imp-1',
+  version: 3,
+  metadata: correlationId === undefined ? {} : { correlationId },
 });
 
-describe('newOperation', () => {
-  it('mints a fresh story and a command causation naming its own command', () => {
-    const context = newOperation(source(STORY, 'command-1'));
+function collecting(): { stream: DestinationStream; lines: string[] } {
+  const lines: string[] = [];
+  return { lines, stream: { write: (chunk: string) => void lines.push(chunk) } };
+}
 
-    expect(context).toEqual({
-      correlationId: STORY,
-      causation: { kind: 'command', commandId: 'command-1' },
-    });
-  });
+describe("this module's correlation binding", () => {
+  it('namespaces a derived causation with this context, not the other one', () => {
+    // A reference whose context is not this one names a row in a DIFFERENT database; the namespace
+    // is the only thing standing between a future resolver and the wrong store.
+    const { context } = continueFrom(triggeringEvent(STORY), fixedCorrelation(OTHER_STORY));
 
-  it('mints a distinct story per unit of work', () => {
-    const minter = source(STORY, 'command-1', OTHER, 'command-2');
-
-    expect(newOperation(minter).correlationId).not.toBe(newOperation(minter).correlationId);
-  });
-});
-
-describe('continueFrom', () => {
-  it('copies the triggering event story verbatim and reports it as carried', () => {
-    const { context, origin } = continueFrom(
-      storedEvent({
-        importId: 'imp-1',
-        occurredAt: '2026-08-06T00:00:00.000Z',
-        correlationId: STORY,
-      }),
-      source(OTHER, 'unused'),
-    );
-
-    expect(origin).toBe('carried');
-    expect(context).toEqual({
-      correlationId: STORY,
-      causation: { kind: 'event', context: CONTEXT_NAME, streamId: 'imp-1', version: 3 },
-    });
-  });
-
-  it('mints a fresh story when the triggering event predates correlation metadata, reporting it as absent', () => {
-    // `absent` and `malformed` are not interchangeable — a pre-capability row is permanent and
-    // unactionable, an unusable one means a writer is emitting bad ids right now — and the caller
-    // logs them at different levels off this field alone.
-    const { context, origin } = continueFrom(
-      storedEvent({ importId: 'imp-1', occurredAt: '2026-08-06T00:00:00.000Z' }),
-      source(OTHER),
-    );
-
-    expect(origin).toBe('absent');
-    expect(context).toEqual({
-      correlationId: OTHER,
-      causation: { kind: 'event', context: CONTEXT_NAME, streamId: 'imp-1', version: 3 },
-    });
-  });
-
-  it('mints a fresh story when the stored id is not a well-formed correlation id, reporting it as malformed', () => {
-    const { context, origin } = continueFrom(
-      storedEvent({
-        importId: 'imp-1',
-        occurredAt: '2026-08-06T00:00:00.000Z',
-        correlationId: 'not-a-trace-id',
-      }),
-      source(OTHER),
-    );
-
-    expect(origin).toBe('malformed');
-    expect(context.correlationId).toBe(OTHER);
-  });
-});
-
-describe('adoptStory', () => {
-  it('adopts a story minted elsewhere verbatim under the given causation', () => {
-    const causation = causedBy('downloader', 'acq-9', 4);
-
-    expect(adoptStory(toCorrelationId(STORY), causation)).toEqual({
-      correlationId: STORY,
-      causation: { kind: 'event', context: 'downloader', streamId: 'acq-9', version: 4 },
-    });
-  });
-});
-
-describe('parseCausation', () => {
-  it('accepts a stored event reference', () => {
-    expect(parseCausation({ kind: 'event', context: 'x', streamId: 's', version: 3 })).toEqual({
+    expect(CONTEXT_NAME).toBe('importer');
+    expect(context.causation).toEqual({
       kind: 'event',
-      context: 'x',
-      streamId: 's',
+      context: 'importer',
+      streamId: 'imp-1',
       version: 3,
     });
   });
 
-  it('accepts version zero — the first event of a stream is a legitimate parent', () => {
-    expect(parseCausation({ kind: 'event', context: 'x', streamId: 's', version: 0 })).toEqual({
-      kind: 'event',
-      context: 'x',
-      streamId: 's',
-      version: 0,
-    });
+  it('still degrades a story it cannot use, through the shared rule', () => {
+    // One case is enough to prove the binding really delegates; the full matrix lives in eventing.
+    const { context, origin } = continueFrom(triggeringEvent(), fixedCorrelation(OTHER_STORY));
+
+    expect(origin).toBe('absent');
+    expect(context.correlationId).toBe(OTHER_STORY);
   });
 
-  it('accepts a stored command reference', () => {
-    expect(parseCausation({ kind: 'command', commandId: 'c-1' })).toEqual({
-      kind: 'command',
-      commandId: 'c-1',
-    });
-  });
+  it('binds the story onto this module’s own logger type', () => {
+    const { stream, lines } = collecting();
+    const logger = createLogger({ level: 'info', destination: stream });
+    const context = newOperation(fixedCorrelation(STORY, 'command-1'));
 
-  it.each([
-    ['a null', null],
-    // The column is absent altogether on rows written before causation existed, so `undefined` is
-    // the likeliest input of all — and the one a reader must be sure cannot fault the parse.
-    ['an absent value', undefined],
-    ['a primitive', 'event'],
-    ['an unknown kind', { kind: 'saga', streamId: 's', version: 1 }],
-    ['a missing kind', { context: 'x', streamId: 's', version: 1 }],
-    ['an empty commandId', { kind: 'command', commandId: '' }],
-    ['a missing streamId', { kind: 'event', context: 'x', version: 1 }],
-    ['an empty context', { kind: 'event', context: '', streamId: 's', version: 1 }],
-    ['an empty streamId', { kind: 'event', context: 'x', streamId: '', version: 1 }],
-    ['a non-string streamId', { kind: 'event', context: 'x', streamId: 7, version: 1 }],
-    ['a non-string context', { kind: 'event', context: 7, streamId: 's', version: 1 }],
-    ['a non-string commandId', { kind: 'command', commandId: 7 }],
-    ['a negative version', { kind: 'event', context: 'x', streamId: 's', version: -1 }],
-    ['a fractional version', { kind: 'event', context: 'x', streamId: 's', version: 1.5 }],
-    ['a stringly version', { kind: 'event', context: 'x', streamId: 's', version: '1' }],
-  ])('drops %s rather than handing back a union nothing checked', (_label, value) => {
-    // The column is JSON written by some past version of this process. A reader narrowing on
-    // `kind` would otherwise read `undefined` off a shape TypeScript never verified.
-    expect(parseCausation(value)).toBeUndefined();
-  });
-});
+    const scope = operationScope(context, logger, { importId: 'imp-1' });
+    scope.logger.info('working');
 
-describe('adoptOrMint', () => {
-  it('adopts a well-formed caller story and reports it as carried', () => {
-    // The source can only mint OTHER, so adopting STORY and re-minting are distinguishable.
-    const { context, origin } = adoptOrMint(STORY, source(OTHER));
-
-    expect(origin).toBe('carried');
-    expect(context.correlationId).toBe(STORY);
-    expect(context.causation).toEqual({ kind: 'command', commandId: OTHER });
-  });
-
-  it('mints fresh for an unusable caller story and reports it as malformed', () => {
-    // A live caller got it wrong — unlike a pre-capability row, that IS actionable.
-    const { context, origin } = adoptOrMint('not-a-trace-id', source(OTHER));
-
-    expect(origin).toBe('malformed');
-    expect(context.correlationId).toBe(OTHER);
+    expect(lines.join('')).toContain(STORY);
+    expect(lines.join('')).toContain('imp-1');
   });
 });
