+++ test/boundaries/correlation.test.ts
@@ -80,67 +80,10 @@ describe('no correlation id reaches user-visible copy', () => {
 });
 
 /**
- * D13 duplicates the correlation module into both bounded contexts on purpose: a shared module
- * would be a shared kernel over the one identity that crosses the seam, and the repo already owns
- * byte-identical copies of `EventStorePort` and `EventMetadata` for the same reason.
- *
- * Deliberate duplication still has to be protected, or it drifts into accidental divergence — at
- * which point the two contexts disagree about what a story IS, and the seam's whole premise (each
- * side reads the other's envelope with its own schema) quietly stops holding.
+ * The per-context correlation twins this file used to pin string-for-string are gone: the mechanism
+ * moved to `@music/eventing` (extract-eventing-package D1/D5), so there is no deliberate duplication
+ * left to protect from drift. What remains per context is a thin binding — the context name stamped
+ * onto derived coordinates, and that module's own logger type — and each package's own suite covers
+ * it. The shared-kernel avoidance this block guarded was retired with the rule it served: the line
+ * is now no shared MODEL, and correlation is mechanism.
  */
-describe('the per-context correlation modules are duplicated deliberately, not drifting', () => {
-  /** Erase what each context is ALLOWED to differ on: its own name, and its aggregate's id field. */
-  const normalize = (source: string): string =>
-    source
-      .replaceAll(/\bdownloader\b|\bimporter\b/g, 'CONTEXT')
-      .replaceAll(/\bacquisitionId\b|\bimportId\b/g, 'STREAM_ID')
-      .replaceAll(/\bacq-1\b|\bimp-1\b/g, 'STREAM');
-
-  it.each([
-    ['correlation/context.ts', 'src/application/correlation/context.ts'],
-    ['correlation/correlation-id.ts', 'src/application/correlation/correlation-id.ts'],
-    ['__fixtures__/correlation.ts', 'src/application/__fixtures__/correlation.ts'],
-  ])('%s is identical in both contexts modulo the context name', (_label, relative) => {
-    const read = (module_: string): string =>
-      readFileSync(path.join(REPO_ROOT, 'packages', module_, relative), 'utf8');
-
-    expect(normalize(read('importer'))).toEqual(normalize(read('downloader')));
-  });
-});
-
-/**
- * The story format is defined in THREE places that share no code: each module's
- * `correlation-id.ts`, and the web BFF's mint. That is deliberate — a story belongs to neither
- * module (one request drives both), so the shell that spans them owns its own mint, and making
- * either module's constant the currency would be the shared kernel D13 exists to avoid.
- *
- * The drift mode is the worst in the change and the quietest: if the BFF minted a shape the
- * modules reject, both facades would discard every caller's story as malformed and mint their
- * own, and NOTHING would log it — the symptom is only that web lines and module lines stop
- * joining. So the three are pinned against each other here.
- */
-describe('the story format agrees everywhere it is defined', () => {
-  const MODULE_PATTERN = '/^[0-9a-f]{32}$/';
-
-  it.each([
-    ['downloader', 'packages/downloader/src/application/correlation/correlation-id.ts'],
-    ['importer', 'packages/importer/src/application/correlation/correlation-id.ts'],
-  ])('the %s module declares the expected pattern', (_module, relative) => {
-    const source = readFileSync(path.join(REPO_ROOT, relative), 'utf8');
-
-    expect(source).toContain(`CORRELATION_ID_PATTERN = ${MODULE_PATTERN}`);
-  });
-
-  it('the BFF mints a value both modules would accept', () => {
-    // Behavioural, not textual: the BFF builds the id from bytes rather than declaring a pattern,
-    // so the only honest pin is that what it produces satisfies the modules' own predicate.
-    const source = readFileSync(
-      path.join(REPO_ROOT, 'packages/web/src/lib/server/correlation.ts'),
-      'utf8',
-    );
-    const bytes = /STORY_BYTES = (\d+)/.exec(source)?.[1];
-
-    expect(bytes).toBe('16'); // 16 bytes hex-encoded = the 32 characters the modules require
-    expect(source).toContain("padStart(2, '0')"); // ...and every byte contributes exactly two
-  });
-});
